requirements/base.txt
asgiref==3.3.1
astroid==2.4.2
Django==3.1.4
django-cors-headers==3.3.0
django-countries==7.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
flake8==3.8.4
greenlet==0.4.17
isort==5.6.4
ipython==7.19.0
ipython-genutils==0.2.0

jedi==0.18.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
neovim==0.3.1
parso==0.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pynvim==0.4.2
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
user-authentication-jr==0.0.7
wrapt==1.12.1